Add files via upload
the new class contain the purchase method read text file list of name, and find a text file, named after the customer in the initial text file, that contain the items and quantities and output all of them into a list of sales. the first 2 methods are file handled to update the stock list in the retailinventory class. the last method is to implement those 2 method concurrently and updates it. I reimplement the savetofile  from class retailinventory to it since the method is privated, care for the potential overrides if you gonna merge the new class to retailinventory.
you can remove the purchase method, and just merge a sales text file/data to the inventory data and pull from it